+++ src/main/java/com/theocean/fundering/domain/celebrity/service/CelebService.java
@@ -3,6 +3,7 @@ package com.theocean.fundering.domain.celebrity.service;
 import com.theocean.fundering.domain.celebrity.domain.Celebrity;
 import com.theocean.fundering.domain.celebrity.dto.*;
 import com.theocean.fundering.domain.celebrity.repository.CelebRepository;
+import com.theocean.fundering.global.dto.PageResponse;
 import com.theocean.fundering.global.errors.exception.Exception400;
 import com.theocean.fundering.global.errors.exception.Exception500;
 import lombok.RequiredArgsConstructor;
